fix(worker): move AgentRuntimeError to a workflow-safe module to unbreak builds
